Browser/requirements.txt
backports.cached-property >= 1.0.0
grpcio == 1.35.0
grpcio-tools == 1.35.0
protobuf==3.14.0
robotframework >=3.2.2
robotframework-pythonlibcore >=2.2.1
typing-extensions >= 3.7.4.3
wrapt >= 1.12.1
